+++ eds-aic-20230930/alibabacloud_eds_aic20230930/client.py
@@ -2528,6 +2528,8 @@ class Client(OpenApiClient):
             query['AndroidInstanceIds'] = request.android_instance_ids
         if not UtilClient.is_unset(request.android_instance_name):
             query['AndroidInstanceName'] = request.android_instance_name
+        if not UtilClient.is_unset(request.authorized_user_id):
+            query['AuthorizedUserId'] = request.authorized_user_id
         if not UtilClient.is_unset(request.biz_region_id):
             query['BizRegionId'] = request.biz_region_id
         if not UtilClient.is_unset(request.charge_type):
@@ -2593,6 +2595,8 @@ class Client(OpenApiClient):
             query['AndroidInstanceIds'] = request.android_instance_ids
         if not UtilClient.is_unset(request.android_instance_name):
             query['AndroidInstanceName'] = request.android_instance_name
+        if not UtilClient.is_unset(request.authorized_user_id):
+            query['AuthorizedUserId'] = request.authorized_user_id
         if not UtilClient.is_unset(request.biz_region_id):
             query['BizRegionId'] = request.biz_region_id
         if not UtilClient.is_unset(request.charge_type):